+++ src/main.js
@@ -1,5 +1,5 @@
-import MoviePresenter from './presenter/movie-presenter.js';
-import NavigationPresenter from './presenter/navigation-presener.js';
+import FilmPresenter from './presenter/film-presenter.js';
+import NavigationPresenter from './presenter/navigation-presenter.js';
 import FilmsModel from './model/films-model.js';
 import FilterModel from './model/filter-model.js';
 import CommentsModel from './model/comments-model.js';
@@ -7,16 +7,14 @@ import { NavigationItemsType } from './constants.js';
 import ApiService from './api-service.js';
 import StatisticsView from './view/statistics/statistics-view.js';
 import { render, RenderPosition } from './utils/render.js';
+import { Api } from './constants.js';
 
-const AUTHORIZATION = 'adg7723bcls';
-const END_POINT = 'https://16.ecmascript.pages.academy/cinemaddict';
-
-const apiService = new ApiService(END_POINT, AUTHORIZATION);
+const apiService = new ApiService(Api.END_POINT, Api.AUTHORIZATION_TOKEN);
 const filmsModel = new FilmsModel(apiService);
 const commentsModel = new CommentsModel(apiService);
 const filterModel = new FilterModel();
 
-const moviePresenter = new MoviePresenter(filmsModel, commentsModel, filterModel);
+const moviePresenter = new FilmPresenter(filmsModel, commentsModel, filterModel);
 
 let statisticsView = null;
