+++ src/model/films-model.js
@@ -46,7 +46,6 @@ export default class FilmsModel extends AbstractObservable {
 
       this.notify(updateType, updatedFilm);
     } catch (err) {
-      console.log(err);
       throw new Error('Can\'t update film!');
     }
   }
